post-experiments/app-labelling/requirements.txt
numpy
opencv-python
pillow
pymongo
python-dotenv
streamlit